[bazel] Separate test individual tests code from helper code for testing.
We needed to depend on `scala_test` target and thus bring in tons of
individual test code to get a minor amount of helper code for tests.
Rearranged code so that we can define separate package for the helper
code and only use that as dependencies.

Also made all `scala_test` targets only package visible to reduce
likelihood of people importing them.
