+++ joern-cli/frontends/c2cpg/BUILD
@@ -16,15 +16,15 @@ scala_library(
         "//bazel/reexports:io_shiftleft_codepropertygraph_with_deps",
         "//joern-cli/frontends/x2cpg",
         "@maven//:com_google_protobuf_protobuf_java",
+        "@maven//:com_lihaoyi_ujson_3",
         "@maven//:io_joern_eclipse_cdt_core",
-        "@maven//:org_eclipse_platform_org_eclipse_equinox_common",
+        "@maven//:org_apache_commons_commons_lang3",
         "@maven//:org_eclipse_platform_org_eclipse_core_resources",
         "@maven//:org_eclipse_platform_org_eclipse_core_runtime",
-        "@maven//:org_eclipse_platform_org_eclipse_text",
+        "@maven//:org_eclipse_platform_org_eclipse_equinox_common",
         "@maven//:org_eclipse_platform_org_eclipse_osgi",
-        "@maven//:org_apache_commons_commons_lang3",
+        "@maven//:org_eclipse_platform_org_eclipse_text",
         "@maven//:org_jline_jline",
-        "@maven//:com_lihaoyi_ujson_3",
     ],
 )
 
@@ -39,8 +39,7 @@ scala_test(
     deps = [
         ":c2cpg",
         "//bazel/reexports:io_shiftleft_codepropertygraph_with_deps",
-        "//dataflowengineoss:tests",
-        "//joern-cli/frontends/x2cpg:tests",
+        "//dataflowengineoss/src/test/scala/io/joern/dataflowengineoss/testfixtures",
         "//semanticcpg",
         "@maven//:com_github_scopt_scopt_3",
         "@maven//:org_scala_lang_modules_scala_parallel_collections_3",
