+++ joern-cli/frontends/javasrc2cpg/src/test/scala/io/joern/javasrc2cpg/querying/RemoteDependencyTests.scala
@@ -1,9 +1,10 @@
 package io.joern.javasrc2cpg.querying
 
 import io.joern.javasrc2cpg.testfixtures.JavaSrcCode2CpgFixture
-import io.joern.x2cpg.utils.ArtifactFetcherTests.artifactFetcherTest
-import io.joern.x2cpg.utils.RetryableAssertion.eventually
-import io.joern.x2cpg.utils.{HttpArtifact, NetworkTest}
+import io.joern.x2cpg.testfixtures.RetryableAssertion.eventually
+import io.joern.x2cpg.testfixtures.NetworkTest
+import io.joern.x2cpg.utils.HttpArtifact
+
 import io.shiftleft.semanticcpg.language.*
 import org.scalatest.Assertion
