+++ joern-cli/frontends/jssrc2cpg/BUILD
@@ -46,9 +46,9 @@ scala_test(
     deps = [
         ":jssrc2cpg",
         "//bazel/reexports:io_shiftleft_codepropertygraph_with_deps",
-        "//dataflowengineoss",
-        "//dataflowengineoss:tests",
-        "//joern-cli/frontends/x2cpg:tests",
+        "//dataflowengineoss/src/test/scala/io/joern/dataflowengineoss/testfixtures",
+        "//semanticcpg",
+        "@maven//:org_scala_lang_modules_scala_parallel_collections_3",
     ],
 )
