+++ joern-cli/frontends/kotlin2cpg/BUILD
@@ -42,9 +42,9 @@ scala_test(
     deps = [
         ":kotlin2cpg",
         "//bazel/reexports:io_shiftleft_codepropertygraph_with_deps",
-        "//dataflowengineoss",
-        "//dataflowengineoss:tests",
-        "//joern-cli/frontends/x2cpg:tests",
+        "//dataflowengineoss/src/test/scala/io/joern/dataflowengineoss/testfixtures",
+        "//semanticcpg",
+        "@maven//:org_scala_lang_modules_scala_parallel_collections_3",
     ],
 )
