+++ joern-cli/frontends/x2cpg/src/test/scala/io/joern/x2cpg/testfixtures/NetworkTest.scala
@@ -1,4 +1,4 @@
-package io.joern.x2cpg.utils
+package io.joern.x2cpg.testfixtures
 
 import org.scalatest.Tag
